Sanitized instance ID to integer

+++ lib/Utils/AMQHandler.php
@@ -93,7 +93,7 @@ class AMQHandler extends Stomp {
         $this->connect();
         $this->setReadTimeout( 5, 0 );
         $this->queueName = $queueName;
-        return parent::subscribe( '/queue/' . INIT::$INSTANCE_ID . "_" . $queueName );
+        return parent::subscribe( '/queue/' . (int)INIT::$INSTANCE_ID . "_" . $queueName );
 
     }
 
@@ -115,7 +115,7 @@ class AMQHandler extends Stomp {
         }
 
         $this->clientType = self::CLIENT_TYPE_PUBLISHER;
-        return parent::send( '/queue/' . INIT::$INSTANCE_ID . "_" . $destination, $msg, $properties, $sync );
+        return parent::send( '/queue/' . (int)INIT::$INSTANCE_ID . "_" . $destination, $msg, $properties, $sync );
 
     }
 
@@ -137,7 +137,7 @@ class AMQHandler extends Stomp {
             throw new \Exception( 'No queue name provided.' );
         }
 
-        $queue_interface_url = INIT::$QUEUE_JMX_ADDRESS . "/api/jolokia/read/org.apache.activemq:type=Broker,brokerName=localhost,destinationType=Queue,destinationName=" . INIT::$INSTANCE_ID . "_" . $queue . "/QueueSize";
+        $queue_interface_url = INIT::$QUEUE_JMX_ADDRESS . "/api/jolokia/read/org.apache.activemq:type=Broker,brokerName=localhost,destinationType=Queue,destinationName=" . (int) INIT::$INSTANCE_ID . "_" . $queue . "/QueueSize";
 
         $mHandler = new MultiCurlHandler();
 
@@ -180,7 +180,7 @@ class AMQHandler extends Stomp {
             throw new \Exception( 'No queue name provided.' );
         }
 
-        $queue_interface_url = INIT::$QUEUE_JMX_ADDRESS . "/api/jolokia/read/org.apache.activemq:type=Broker,brokerName=localhost,destinationType=Queue,destinationName=" . INIT::$INSTANCE_ID . "_" . $queue . "/ConsumerCount";
+        $queue_interface_url = INIT::$QUEUE_JMX_ADDRESS . "/api/jolokia/read/org.apache.activemq:type=Broker,brokerName=localhost,destinationType=Queue,destinationName=" . (int) INIT::$INSTANCE_ID . "_" . $queue . "/ConsumerCount";
 
         $mHandler = new MultiCurlHandler();
