+++ lib/Utils/Analysis/Workers/FastAnalysis.php
@@ -100,7 +100,7 @@ class FastAnalysis extends AbstractDaemon {
 
         try {
             self::$queueHandler = new AMQHandler();
-            self::$queueHandler->getRedisClient()->sadd( RedisKeys::FAST_PID_LIST, self::$tHandlerPID . ":" . gethostname() . ":" . INIT::$INSTANCE_ID );
+            self::$queueHandler->getRedisClient()->sadd( RedisKeys::FAST_PID_LIST, self::$tHandlerPID . ":" . gethostname() . ":" . (int) INIT::$INSTANCE_ID );
 
             $this->_updateConfiguration();
 
@@ -343,7 +343,7 @@ class FastAnalysis extends AbstractDaemon {
         self::$tHandlerPID = null;
 
         //SHUTDOWN
-        self::$queueHandler->getRedisClient()->srem( RedisKeys::FAST_PID_LIST, getmypid() . ":" . gethostname() . ":" . INIT::$INSTANCE_ID );
+        self::$queueHandler->getRedisClient()->srem( RedisKeys::FAST_PID_LIST, getmypid() . ":" . gethostname() . ":" . (int) INIT::$INSTANCE_ID );
 
         $msg = str_pad( " FAST ANALYSIS " . getmypid() . ":" . gethostname() . ":" . INIT::$INSTANCE_ID . " HALTED GRACEFULLY ", 50, "-", STR_PAD_BOTH );
         self::_TimeStampMsg( $msg );
@@ -750,7 +750,7 @@ HD;
         $and_InstanceId = null;
         if( !is_null( INIT::$INSTANCE_ID ) ){
             $and_InstanceId = ' AND instance_id = :instance_id ';
-            $bindParams[ 'instance_id' ] = INIT::$INSTANCE_ID;
+            $bindParams[ 'instance_id' ] = (int) INIT::$INSTANCE_ID;
         }
 
         $query = "
